fix(B-464): split app config model and schema
