+++ backend/app/schemas/__init__.py
@@ -1,4 +1,11 @@
 from app.schemas.admin_dashboard_stats import AdminDashboardStatsRead
+from app.schemas.app_config import AppConfigCreate, AppConfigRead
 from app.schemas.message_schema import MessageCreateSchema, MessageSchema
 
-__all__ = ['MessageSchema', 'MessageCreateSchema', 'AdminDashboardStatsRead']
+__all__ = [
+    'MessageSchema',
+    'MessageCreateSchema',
+    'AdminDashboardStatsRead',
+    'AppConfigCreate',
+    'AppConfigRead',
+]
